add limited plane

+++ src/colors_utils.c
@@ -6,7 +6,7 @@
 /*   By: abenaiss <abenaiss@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/11/19 18:58:42 by abenaiss          #+#    #+#             */
-/*   Updated: 2020/01/14 18:27:04 by abenaiss         ###   ########.fr       */
+/*   Updated: 2020/01/19 21:06:45 by abenaiss         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,17 +26,17 @@ int			ft_rgb_to_int(t_color color)
 	return (ft_clip_min(0, c));
 }
 
-double		ft_clip_min(int min, double value)
+double		ft_clip_min(double min, double value)
 {
 	return ((value >= min) ? value : min);
 }
 
-double		ft_clip_max(int max, double value)
+double		ft_clip_max(double max, double value)
 {
 	return ((value <= max) ? value : max);
 }
 
-double		ft_clip_min_max(int min, int max, double value)
+double		ft_clip_min_max(double min, double max, double value)
 {
 	return (ft_clip_max(max, ft_clip_min(min, value)));
 }
